feat: 작품관리 ui 수정 (요일 필터링 드롭다운 + 전체 작품보기 버튼))
